+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysRoleMenuMapper.java
@@ -1,8 +1,8 @@
 package com.ruoyi.system.mapper;
 
-import java.util.List;
-
 import com.ruoyi.system.domain.SysRoleMenu;
+
+import java.util.List;
 
 /**
  * 角色与菜单关联表 数据层
@@ -16,7 +16,7 @@ public interface SysRoleMenuMapper {
      * @param menuId 菜单ID
      * @return 结果
      */
-    int checkMenuExistRole(Long menuId);
+    Long checkMenuExistRole(Long menuId);
 
     /**
      * 通过角色ID删除角色和菜单关联
@@ -24,7 +24,7 @@ public interface SysRoleMenuMapper {
      * @param roleId 角色ID
      * @return 结果
      */
-    int deleteRoleMenuByRoleId(Long roleId);
+    Long deleteRoleMenuByRoleId(Long roleId);
 
     /**
      * 批量删除角色菜单关联信息
@@ -32,7 +32,7 @@ public interface SysRoleMenuMapper {
      * @param ids 需要删除的数据ID
      * @return 结果
      */
-    int deleteRoleMenu(Long[] ids);
+    Long deleteRoleMenu(Long[] ids);
 
     /**
      * 批量新增角色菜单信息
@@ -40,5 +40,5 @@ public interface SysRoleMenuMapper {
      * @param roleMenuList 角色菜单列表
      * @return 结果
      */
-    int batchRoleMenu(List<SysRoleMenu> roleMenuList);
+    Long batchRoleMenu(List<SysRoleMenu> roleMenuList);
 }
