+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysUserRoleMapper.java
@@ -1,9 +1,9 @@
 package com.ruoyi.system.mapper;
 
-import java.util.List;
-
-import org.apache.ibatis.annotations.Param;
 import com.ruoyi.system.domain.SysUserRole;
+import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
 
 /**
  * 用户与角色关联表 数据层
@@ -17,7 +17,7 @@ public interface SysUserRoleMapper {
      * @param userId 用户ID
      * @return 结果
      */
-    int deleteUserRoleByUserId(Long userId);
+    Long deleteUserRoleByUserId(Long userId);
 
     /**
      * 批量删除用户和角色关联
@@ -25,7 +25,7 @@ public interface SysUserRoleMapper {
      * @param ids 需要删除的数据ID
      * @return 结果
      */
-    int deleteUserRole(Long[] ids);
+    Long deleteUserRole(Long[] ids);
 
     /**
      * 通过角色ID查询角色使用数量
@@ -33,7 +33,7 @@ public interface SysUserRoleMapper {
      * @param roleId 角色ID
      * @return 结果
      */
-    int countUserRoleByRoleId(Long roleId);
+    Long countUserRoleByRoleId(Long roleId);
 
     /**
      * 批量新增用户角色信息
@@ -41,7 +41,7 @@ public interface SysUserRoleMapper {
      * @param userRoleList 用户角色列表
      * @return 结果
      */
-    int batchUserRole(List<SysUserRole> userRoleList);
+    Long batchUserRole(List<SysUserRole> userRoleList);
 
     /**
      * 删除用户和角色关联信息
@@ -49,7 +49,7 @@ public interface SysUserRoleMapper {
      * @param userRole 用户和角色关联信息
      * @return 结果
      */
-    int deleteUserRoleInfo(SysUserRole userRole);
+    Long deleteUserRoleInfo(SysUserRole userRole);
 
     /**
      * 批量取消授权用户角色
@@ -58,5 +58,5 @@ public interface SysUserRoleMapper {
      * @param userIds 需要删除的用户数据ID
      * @return 结果
      */
-    int deleteUserRoleInfos(@Param("roleId") Long roleId, @Param("userIds") Long[] userIds);
+    Long deleteUserRoleInfos(@Param("roleId") Long roleId, @Param("userIds") Long[] userIds);
 }
