+++ ruoyi-system/src/main/java/com/ruoyi/system/service/ISysDictDataService.java
@@ -1,8 +1,8 @@
 package com.ruoyi.system.service;
 
-import java.util.List;
-
 import com.ruoyi.common.core.domain.entity.SysDictData;
+
+import java.util.List;
 
 /**
  * 字典 业务层
@@ -48,7 +48,7 @@ public interface ISysDictDataService {
      * @param dictData 字典数据信息
      * @return 结果
      */
-    int insertDictData(SysDictData dictData);
+    Long insertDictData(SysDictData dictData);
 
     /**
      * 修改保存字典数据信息
@@ -56,5 +56,5 @@ public interface ISysDictDataService {
      * @param dictData 字典数据信息
      * @return 结果
      */
-    int updateDictData(SysDictData dictData);
+    Long updateDictData(SysDictData dictData);
 }
