+++ ruoyi-system/src/main/java/com/ruoyi/system/service/ISysLogininforService.java
@@ -1,8 +1,8 @@
 package com.ruoyi.system.service;
 
-import java.util.List;
-
 import com.ruoyi.system.domain.SysLogininfor;
+
+import java.util.List;
 
 /**
  * 系统访问日志情况信息 服务层
@@ -31,7 +31,7 @@ import com.ruoyi.system.domain.SysLogininfor;
      * @param infoIds 需要删除的登录日志ID
      * @return 结果
      */
-     int deleteLogininforByIds(Long[] infoIds);
+     Long deleteLogininforByIds(Long[] infoIds);
 
     /**
      * 清空系统登录日志
