+++ ruoyi-system/src/main/java/com/ruoyi/system/service/ISysMenuService.java
@@ -1,11 +1,11 @@
 package com.ruoyi.system.service;
 
-import java.util.List;
-import java.util.Set;
-
 import com.ruoyi.common.core.domain.TreeSelect;
 import com.ruoyi.common.core.domain.entity.SysMenu;
 import com.ruoyi.system.domain.vo.RouterVo;
+
+import java.util.List;
+import java.util.Set;
 
 /**
  * 菜单 业务层
@@ -116,7 +116,7 @@ import com.ruoyi.system.domain.vo.RouterVo;
      * @param menu 菜单信息
      * @return 结果
      */
-     int insertMenu(SysMenu menu);
+     Long insertMenu(SysMenu menu);
 
     /**
      * 修改保存菜单信息
@@ -124,7 +124,7 @@ import com.ruoyi.system.domain.vo.RouterVo;
      * @param menu 菜单信息
      * @return 结果
      */
-     int updateMenu(SysMenu menu);
+     Long updateMenu(SysMenu menu);
 
     /**
      * 删除菜单管理信息
@@ -132,7 +132,7 @@ import com.ruoyi.system.domain.vo.RouterVo;
      * @param menuId 菜单ID
      * @return 结果
      */
-     int deleteMenuById(Long menuId);
+     Long deleteMenuById(Long menuId);
 
     /**
      * 校验菜单名称是否唯一
