+++ ruoyi-system/src/main/java/com/ruoyi/system/service/ISysOperLogService.java
@@ -1,8 +1,8 @@
 package com.ruoyi.system.service;
 
-import java.util.List;
-
 import com.ruoyi.system.domain.SysOperLog;
+
+import java.util.List;
 
 /**
  * 操作日志 服务层
@@ -31,7 +31,7 @@ import com.ruoyi.system.domain.SysOperLog;
      * @param operIds 需要删除的操作日志ID
      * @return 结果
      */
-     int deleteOperLogByIds(Long[] operIds);
+    Long deleteOperLogByIds(Long[] operIds);
 
     /**
      * 查询操作日志详细
