+++ ruoyi-system/src/main/java/com/ruoyi/system/service/ISysRoleService.java
@@ -1,10 +1,10 @@
 package com.ruoyi.system.service;
 
-import java.util.List;
-import java.util.Set;
-
 import com.ruoyi.common.core.domain.entity.SysRole;
 import com.ruoyi.system.domain.SysUserRole;
+
+import java.util.List;
+import java.util.Set;
 
 /**
  * 角色业务层
@@ -95,7 +95,7 @@ public interface ISysRoleService {
      * @param roleId 角色ID
      * @return 结果
      */
-    int countUserRoleByRoleId(Long roleId);
+    Long countUserRoleByRoleId(Long roleId);
 
     /**
      * 新增保存角色信息
@@ -103,7 +103,7 @@ public interface ISysRoleService {
      * @param role 角色信息
      * @return 结果
      */
-    int insertRole(SysRole role);
+    Long insertRole(SysRole role);
 
     /**
      * 修改保存角色信息
@@ -111,7 +111,7 @@ public interface ISysRoleService {
      * @param role 角色信息
      * @return 结果
      */
-    int updateRole(SysRole role);
+    Long updateRole(SysRole role);
 
     /**
      * 修改角色状态
@@ -119,7 +119,7 @@ public interface ISysRoleService {
      * @param role 角色信息
      * @return 结果
      */
-    int updateRoleStatus(SysRole role);
+    Long updateRoleStatus(SysRole role);
 
     /**
      * 修改数据权限信息
@@ -127,7 +127,7 @@ public interface ISysRoleService {
      * @param role 角色信息
      * @return 结果
      */
-    int authDataScope(SysRole role);
+    Long authDataScope(SysRole role);
 
     /**
      * 通过角色ID删除角色
@@ -135,7 +135,7 @@ public interface ISysRoleService {
      * @param roleId 角色ID
      * @return 结果
      */
-    int deleteRoleById(Long roleId);
+    Long deleteRoleById(Long roleId);
 
     /**
      * 批量删除角色信息
@@ -143,7 +143,7 @@ public interface ISysRoleService {
      * @param roleIds 需要删除的角色ID
      * @return 结果
      */
-    int deleteRoleByIds(Long[] roleIds);
+    Long deleteRoleByIds(Long[] roleIds);
 
     /**
      * 取消授权用户角色
@@ -151,7 +151,7 @@ public interface ISysRoleService {
      * @param userRole 用户和角色关联信息
      * @return 结果
      */
-    int deleteAuthUser(SysUserRole userRole);
+    Long deleteAuthUser(SysUserRole userRole);
 
     /**
      * 批量取消授权用户角色
@@ -160,7 +160,7 @@ public interface ISysRoleService {
      * @param userIds 需要取消授权的用户数据ID
      * @return 结果
      */
-    int deleteAuthUsers(Long roleId, Long[] userIds);
+    Long deleteAuthUsers(Long roleId, Long[] userIds);
 
     /**
      * 批量选择授权用户角色
@@ -169,5 +169,5 @@ public interface ISysRoleService {
      * @param userIds 需要删除的用户数据ID
      * @return 结果
      */
-    int insertAuthUsers(Long roleId, Long[] userIds);
+    Long insertAuthUsers(Long roleId, Long[] userIds);
 }
